# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for a chat application.",
        "title": "Chat App API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/v1/health": {
            "get": {
                "description": "Get the health status of the service",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Health check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/rooms": {
            "get": {
                "description": "List available chat rooms",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chat"
                ],
                "summary": "Get chat rooms",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.UIRoom"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/rooms/{room}/bind": {
            "get": {
                "description": "Bind to a given chat room",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chat"
                ],
                "summary": "Bind to chat room",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Room name",
                        "name": "room",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Nickname",
                        "name": "nickname",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Connected",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/rooms/{room}/send": {
            "get": {
                "description": "Send a message to a specific room identified by room ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "websocket"
                ],
                "summary": "Send a message to a specific room",
                "parameters": [
                    {
                        "type": "string",
                        "description": "room ID",
                        "name": "room",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "nickname",
                        "name": "nickname",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Payload with nickname and message",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Message": {
            "type": "object",
            "required": [
                "content",
                "nickname"
            ],
            "properties": {
                "content": {
                    "type": "string"
                },
                "nickname": {
                    "type": "string"
                },
                "room": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
        "models.UIRoom": {
            "type": "object",
            "properties": {
                "room": {
                    "type": "string"
                },
                "users": {
                    "type": "integer"
                }
            }
        }
    }
}